add gambler skill A
